🚚 Moved resources outside of utils folder
