+++ zabbixci/git/credentials.py
