+++ zabbixci/handlers/synchronization/__init__.py
